adding factory for inferring type from dataset value
