+++ tests/ui_tests/test_ui_transformation_view.py
@@ -4,9 +4,12 @@ import h5py
 from PySide2.QtGui import QVector3D
 from mock import Mock
 
-from nexus_constructor.field_attrs import _get_human_readable_type
 from nexus_constructor.model.component import Component
-from nexus_constructor.model.dataset import Dataset, DatasetMetadata
+from nexus_constructor.model.dataset import (
+    Dataset,
+    DatasetMetadata,
+    _get_human_readable_type,
+)
 from nexus_constructor.model.entry import Instrument
 from nexus_constructor.transformation_view import EditRotation, EditTranslation
 from nexus_constructor.validators import FieldType
